@@ -17,7 +17,6 @@ documentation = "https://wikibaseintegrator.readthedocs.io"
 keywords = ["wikibase", "wikidata", "mediawiki", "sparql"]
 classifiers = [
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -39,7 +38,7 @@ classifiers = [
 Changelog = "https://github.com/LeMyst/WikibaseIntegrator/releases"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 backoff = "^2.2.1"
 mwoauth = "^0.4.0"
 oauthlib = "^3.2.2"
